chore: move power of two code to an exported util package
