+++ app/prepare_proposal.go
@@ -3,6 +3,7 @@ package app
 import (
 	"math"
 
+	"github.com/celestiaorg/celestia-app/pkg/util"
 	"github.com/celestiaorg/celestia-app/x/payment/types"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/x/auth/signing"
@@ -69,7 +70,7 @@ func (app *App) estimateSquareSize(data *core.Data) uint64 {
 
 	totalShareEstimate := txShareEstimate + evdShareEstimate + msgShareEstimate
 	sr := math.Sqrt(float64(totalShareEstimate))
-	estimatedSize := types.NextHighestPowerOf2(uint64(sr))
+	estimatedSize := util.NextHighestPowerOf2(uint64(sr))
 	switch {
 	case estimatedSize > consts.MaxSquareSize:
 		return consts.MaxSquareSize
