+++ x/payment/types/payfordata.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"math/bits"
 
+	"github.com/celestiaorg/celestia-app/pkg/util"
 	"github.com/celestiaorg/rsmt2d"
 	sdkclient "github.com/cosmos/cosmos-sdk/client"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -170,54 +171,13 @@ func powerOf2MountainRange(l, k uint64) []uint64 {
 			output = append(output, k)
 			l = l - k
 		case l < k:
-			p := nextLowestPowerOf2(l)
+			p := util.NextLowestPowerOf2(l)
 			output = append(output, p)
 			l = l - p
 		}
 	}
 
 	return output
-}
-
-// NextHighestPowerOf2 returns the next lowest power of 2 unless the input is a power
-// of two, in which case it returns the input
-func NextHighestPowerOf2(v uint64) uint64 {
-	// keep track of the value to check if its the same later
-	i := v
-
-	// find the next highest power using bit mashing
-	v--
-	v |= v >> 1
-	v |= v >> 2
-	v |= v >> 4
-	v |= v >> 8
-	v |= v >> 16
-	v |= v >> 32
-	v++
-
-	// force the value to the next highest power of two if its the same
-	if v == i {
-		return 2 * v
-	}
-
-	return v
-}
-
-func nextLowestPowerOf2(v uint64) uint64 {
-	c := NextHighestPowerOf2(v)
-	if c == v {
-		return c
-	}
-	return c / 2
-}
-
-// Check if number is power of 2
-func powerOf2(v uint64) bool {
-	if v&(v-1) == 0 && v != 0 {
-		return true
-	} else {
-		return false
-	}
 }
 
 // DelimLen calculates the length of the delimiter for a given message size
